Fixed Slime & Hide Spot Sprite Glitch On Kill
Preventing the Slime from killing while hidden will ensure we do not encounter this bug again. The movement code for player and NPC has been moved to FixedUpdate in order to prevent Time.deltaTime issues. Lastly the NPCSightComponent has had its' update frequency increased to ensure that the slime is actually caught if moves in front of the NPCs view.